test: merge DAG API tests
